[BE] Test : log스키마 수정 & front 에서 로그 보내기 테스트
