+++ backend/src/route/log/controller.ts
@@ -3,6 +3,7 @@ import Log from '../../models/Log';
 
 const createLog = async (req: Request, res: Response): Promise<void> => {
   const log = new Log(req.body);
+  console.log('res : ', res);
   await log.save((err, logInfo) => {
     if (err) return res.json({ success: false, err });
     return res.status(200).json({
