+++ frontend/src/components/Common/Card/MagCard/index.tsx
@@ -16,17 +16,6 @@ type MagMeta = {
   tag: string;
 };
 
-const eventLogHandler = () => {
-  axios
-    .post('http://localhost:8000/api/log', {
-      eventName: 'move_page',
-      parameters: { prev: '', next: '' },
-    })
-    .then(res => {
-      console.log(res.data);
-    });
-};
-
 const enterTitle = title => {
   const pop = title.split(':');
   return pop.map((pa, i) => <p key={i}>{pa}</p>);
@@ -35,7 +24,7 @@ const enterTitle = title => {
 const MagCard = ({ magMetaData: mag }: IMagMetaProps) => {
   return (
     <Container>
-      <BoxItem imgUrl={mag.imgUrl} onClick={eventLogHandler} />
+      <BoxItem imgUrl={mag.imgUrl} />
       <MagTitle>{enterTitle(mag?.title)}</MagTitle>
       <MagDate>{mag?.date}</MagDate>
     </Container>
